Q4 enseash.c
The update_prompt function adds a dynamic prompt that displays the termination status of the last command executed. It analyzes the status returned by wait(), WIFEXITED + WEXITSTATUS for a normal exit code (enseash [exit:0] %) and WIFSIGNALED + WTERMSIG for a signal (enseash [sign:2] % after Ctrl+C), or a default prompt otherwise. It is called after each wait(&status). We know whether the function succeeded or not.

Translated with DeepL.com (free version)