Q5 enseash.c
A measure of the execution time of each command has been added to the micro-shell. Just before launching the child process, the program stores the start time, then, just after waiting for the child to finish, it stores the end time and calculates the duration in milliseconds. This duration is then passed to the function that constructs the prompt.