Q6 enseash.c
First, we create a cut_cmd function that cuts the command typed as an argument (argv[0], argv[1]... up to the last one where we have argv[INPUT_BUFSIZE/2]=NULL), then we implement this function in enseash. First, we perform cut_cmd(stock, argv); then, in the condition pid==0, we replace execlp (which cannot be used to execute complex commands) with execvp, which executes the complex command by taking each argument from the split command.